refactor(redis): 重构 Redis缓存操作
- 删除 RedisCache 类，改为使用 RedisKeyCache- 更新相关类中的 RedisCache引用为 RedisKeyCache
- 调整 RedisKeyCache 使用泛型以提高类型安全
- 移除 OnlineUserController 中未使用的代码
- 重命名 CaptchaController 中的 verifyKey 以提高可读性
- 优化 IPUtils 中的 IP 验证逻辑
- 新增 DeviceType 枚举和 BrowserDevice 类以支持设备信息
- 添加 LoginDeviceDTO 和 LoginSessionDTO 类以处理登录相关数据
- 更新应用配置，将 single-login 设置为 false
-优化 AuthService 中的登录逻辑，支持记住我功能
- 更新 RedisConstants 类，添加新的 Redis key常量

+++ chuang-common/chuang-common-core/src/main/java/cn/zhangchuangla/common/core/entity/security/AuthTokenVo.java
@@ -40,6 +40,6 @@ public class AuthTokenVo implements Serializable {
      * 过期时间（单位：秒）
      */
     @Schema(description = "过期时间(单位：秒)")
-    private Integer expires;
+    private Long expires;
 
 }
